feat(server): add local MinIO support for development environment
- Add minio_server dependency for development environment
- Configure MinIO to run on port 9095 with local credentials
- Add MinioBucketCreator to automatically create development bucket
- Add dev_use_remote_storage?() function to switch between local and remote storage
- Configure S3 functions to return MinIO values in development
- Support TUIST_DEV_USE_REMOTE_STORAGE env var to force remote storage in dev
- Update runtime config to handle dynamic scheme and port from endpoint URL
- Set virtual_host to false when using MinIO in development
- Add minio/ directory to .gitignore for local development data
