+++ server/.gitignore
@@ -57,6 +57,7 @@ uuid
 # Local development
 .opencode/
 .claude/
+minio/
 
 # Gettext
 # NOTE: We currently do not utilise gettext for actual translations and relying on key-as-value gettext calls.
